+++ src/Models/WhatsAppMessage.php
@@ -7,6 +7,9 @@ namespace JigarDhulla\LaravelWhatsApp\Models;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Carbon;
 use JigarDhulla\LaravelWhatsApp\Services\WhatsAppMessageReader;
+use Laravel\Ai\Files\Document;
+use Laravel\Ai\Files\File;
+use Laravel\Ai\Files\Image;
 
 /**
  * @property int $rowid
@@ -33,6 +36,10 @@ use JigarDhulla\LaravelWhatsApp\Services\WhatsAppMessageReader;
  */
 class WhatsAppMessage extends Model
 {
+    public const string MIME_TYPE_JPEG = 'image/jpeg';
+
+    public const string MIME_TYPE_PDF = 'application/pdf';
+
     protected $connection = WhatsAppMessageReader::CONNECTION_NAME;
 
     protected $table = 'messages';
@@ -48,5 +55,21 @@ class WhatsAppMessage extends Model
             'ts' => 'timestamp',
             'downloaded_at' => 'timestamp',
         ];
+    }
+
+    /**
+     * @return array|File[]
+     */
+    public function getAttachments(): array
+    {
+        if (is_null($this->media_type) || is_null($this->local_path) || ! file_exists($this->local_path)) {
+            return [];
+        }
+
+        return match ($this->mime_type) {
+            self::MIME_TYPE_JPEG => [Image::fromPath($this->local_path)],
+            self::MIME_TYPE_PDF => [Document::fromPath($this->local_path)],
+            default => []
+        };
     }
 }